YinSearchView
I thought i overrode the get_results method, but its not working and i
don't even think it's calling my method.

on the yinsearch.html file, i basically copied that directly from
search.html.  the only change is i removed the "if query" statement,
since i want it to show results regardless if there is a query (like
when someone hits this page, it should automatically show all the
restaurants, then the results shoul ideally be dynamically filtered as
they type in parameters in the search form).